start adding clean docstring for scrappers and models
