--- dabeplech/models/kegg/__init__.py
@@ -1,3 +1,0 @@
-from .link import KeggLinksModel  # noqa
-from .orthology import KeggOrthologyModel, KeggOrthologyListModel  # noqa
-from .pathway import KeggPathwayModel, KeggPathwayListModel  # noqa
