--- dabeplech/models/kegg/dblinks.py
@@ -1,9 +1,0 @@
-from typing import List
-
-from pydantic import BaseModel
-
-
-class DBLinksModel(BaseModel):
-    GO: List[str] = None
-    COG: List[str] = None
-    RN: List[str] = None
