--- dabeplech/models/kegg/link.py
@@ -1,7 +1,0 @@
-from typing import List, Dict
-
-from pydantic import BaseModel
-
-
-class KeggLinksModel(BaseModel):
-    links: Dict[str, List[str]]
