--- dabeplech/models/ncbi_taxonomy/__init__.py
@@ -1,1 +1,0 @@
-from .taxonomy import NCBITaxonomyAPIModel  # noqa
