--- dabeplech/scrappers/ncbi_taxonomy/__init__.py
@@ -1,1 +1,0 @@
-from .taxonomy import NCBITaxonomyScrapper  # noqa
