+++ docs/api_docs/api_services.rst
@@ -12,13 +12,6 @@ KEGG
 
     KEGGAPI
 
-.. automodule:: dabeplech.kegg
-    :members:
-    :undoc-members:
-    :special-members: __init__
-    :show-inheritance:
-
-
 TogoWS
 ======
