+++ docs/api_docs/scrappers.rst
@@ -1,0 +1,15 @@
+*********
+Scrappers
+*********
+
+Different available scrappers.
+
+NCBI taxonomy
+=============
+
+.. currentmodule:: dabeplech.scrappers.ncbi_taxonomy
+
+.. autosummary::
+    :toctree: stubs
+
+    NCBITaxonomyScrapper
