+++ setup.cfg
@@ -1,0 +1,2 @@
+[pydocstyle]
+match=(?!__init__\.py).*\.py
